+++ Package.swift
@@ -12,14 +12,15 @@ let package = Package(
             targets: ["Indice.Identity"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/indice-co/Indice.Swift.Networking", exact: "1.3.0"),
+        // .package(url: "https://github.com/indice-co/Indice.Swift.Networking", .upToNextMinor(from: "1.4.0")),
+        .package(url: "https://github.com/indice-co/Indice.Swift.Networking", branch: "develop"),
         .package(url: "https://github.com/SwiftyJSON/SwiftyJSON", .upToNextMajor(from: "5.0.2"))
     ],
     targets: [
         .target(
             name: "Indice.Identity",
             dependencies: [
-                .product(name: "IndiceNetworkClient", package: "Indice.Swift.Networking"),
+                .product(name: "NetworkUtilities", package: "Indice.Swift.Networking"),
                 .product(name: "SwiftyJSON", package: "SwiftyJSON")
             ]
         ),