+++ Sources/Indice.Identity/Client/IdentityClientImplementation.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import IndiceNetworkClient
 
 private class Repositories {
     
@@ -70,13 +69,22 @@ internal class IdentityClientImpl: IdentityClient {
     
     public var tokens: TokenStorageAccessor { get { tokenStorage } }
 
-    private let createNetworkClient: (IdentityClient) -> NetworkClient
+    private let networkOptionsBuilder: (IdentityClient) -> NetworkOptions
+    private lazy var networkOptions: NetworkOptions = networkOptionsBuilder(self)
     
-    public
-    private(set) lazy
-    var networkClient: RequestProcessor = {
-        self.createNetworkClient(self)
+    private(set)
+    lazy var requestProcessor: any RequestProcessor = {
+        networkOptions.processor()
     }()
+    
+    private(set)
+    lazy var networkClient: RequestProcessor = {
+        RequestProcessorWrapper(processor: requestProcessor,
+                                tokenAccessor: tokenStorage)
+    }()
+    
+    private(set)
+    lazy var errorParser: ErrorParser = networkOptions.errorParser
     
     private lazy
     var repositories: Repositories = {
@@ -128,7 +136,8 @@ internal class IdentityClientImpl: IdentityClient {
     public
     private(set)
     lazy var userRegistrationService: UserRegistrationService = {
-        UserRegistrationServiceImpl(accountRepository: repositories.accountRepository)
+        UserRegistrationServiceImpl(accountRepository: repositories.accountRepository,
+                                    errorParser: errorParser)
     }()
     
     
@@ -139,17 +148,14 @@ internal class IdentityClientImpl: IdentityClient {
                 currentDeviceInfoProvider: CurrentDeviceInfoProvider,
                 valueStorage: ValueStorage = UserDefaults.standard,
                 tokenStorage: TokenStorage = .ephemeral,
-                networkClientBuilder: ((IdentityClient) -> NetworkClient)? = nil) {
+                networkOptionsBuilder: @escaping ((IdentityClient) -> NetworkOptions)) {
         self.client = client
         self.configuration = configuration
         self.currentDeviceInfoProvider = currentDeviceInfoProvider
         self.valueStorage = valueStorage
         self.tokenStorage = tokenStorage
         self.options = options
-        self.createNetworkClient = networkClientBuilder ?? { client in
-            NetworkClient(interceptors: [AuthorizationHeaderInterceptor(tokenAccessor: client.tokens),
-                                         AuthorizingInterceptor(authServiceProvider: { [weak client] in client?.authorizationService })])
-        }
+        self.networkOptionsBuilder = networkOptionsBuilder
     }
     
 }