+++ Sources/Indice.Identity/Models/Responses/ExtendedProblemDetails.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 import SwiftyJSON
-import IndiceNetworkClient
+
 
 public struct ExtendedProblemDetails: Decodable {